+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 * Remove `EveOnline::XML::CharacterBookmarks`. Use `EveOnline::ESI::CharacterBookmarks` instead.
 * Add `EveOnline::ESI::Models::BookmarkFolder`
 * Add `EveOnline::ESI::CharacterBookmarkFolders`
+* Add `EveOnline::ESI::Models::Notification`
+* Add `EveOnline::ESI::CharacterNotifications`
+* Remove `EveOnline::XML::CharacterContactNotifications`. Use `EveOnline::ESI::CharacterNotifications` instead.
+* Remove `EveOnline::ContactNotification`
 
 **v0.14.0**